contact page component removed
